FigS5
db-RDA ordination plot representing location of deployments across dbRDA axes. Arrows represent the coefficients of each variable on the constrained dbRDA axes, and crosses and dashed lines represent the centroid of factors in the ordination. The length of the arrow and dashed lines represent the strength of the effect of a variable or factor on community composition, and the direction (equal or opposite) indicates whether variables have positive or negative effects on the variation represented by constrained axes. See also Table S8 containing coefficients and centroids of each variable in the dbRDA ordination.