Table S10
Variable coefficients for each of the constrained dbRDA axes and centroids of factor variables on each of the constrained axes. Constrained dbRDA axes are orthogonal combinations of the explanatory variables (i.e., a multiple regression model) that best explain, in successive order, the variation of the response matrix (Borcard, Gillet & Legendre, 2011).