Table S12
SIMPER pairwise comparisons of islands, with mean MaxN (and SD), and average contribution of each species to overall dissimilarities amongst pairs. Significance of the contribution of each species to overall dissimilarities between pairs was assessed using permutation tests. To compare the Tinhosas to Príncipe and São Tomé, only rocky reefs were considered, since the Tinhosas are a rocky reef environment and sand and maerl are not present.